requirements.txt
asttokens==2.4.1
colorama==0.4.6
executing==2.0.1
icecream==2.1.3
Pygments==2.17.2
pyminizip==0.2.6
python-dotenv==1.0.0
pyzip==0.2.0
six==1.16.0
